@@ -4,66 +4,63 @@ description:
   including pinging specific IP address or hostname, sending Wake on LAN magic packets,
   whois and DNS Lookup.
 publish_to: "none"
-version: 2.1.2+9
+version: 2.2.0
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
-  adapty_flutter: ^2.2.5
+  adapty_flutter: ^2.4.4
   animate_do: ^3.0.2
-  bloc: ^8.1.1
-  carrier_info: 2.0.5
+  bloc: ^8.1.2
+  carrier_info: ^2.0.8
   cupertino_icons: ^1.0.5
   cupertino_onboarding: ^1.1.1
-  dart_ping: ^7.0.1
-  dart_ping_ios: ^2.0.0
-  dynamic_color: ^1.6.2
+  dart_ping: ^8.0.1
+  dart_ping_ios: ^3.0.0
+  dynamic_color: ^1.4.0
   enum_to_string: ^2.0.1
   equatable: ^2.0.5
-  flex_color_scheme: ^6.1.2
+  feedback: ^2.6.0
+  feedback_sentry: ^2.3.0
+  flex_color_scheme: ^7.2.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.2
-  flutter_platform_widgets: ^2.0.0
-  flutter_svg: ^1.1.5
-  font_awesome_flutter: ^10.3.0
-  google_fonts: ^4.0.3
-  google_maps_flutter: ^2.2.3
-  google_mobile_ads: ^2.3.0
+  flutter_bloc: ^8.1.3
+  flutter_platform_widgets: ^3.3.5
+  flutter_svg: ^2.0.7
+  font_awesome_flutter: ^10.5.0
+  google_fonts: ^5.1.0
+  google_maps_flutter: ^2.4.0
+  google_mobile_ads: ^3.0.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  http: ^0.13.5
-  hydrated_bloc: ^8.1.0
-  in_app_review: ^2.0.4
-  introduction_screen: ^3.0.2
-  json_annotation: ^4.7.0
-  json_serializable: ^6.5.0
+  http: ^1.0.0
+  hydrated_bloc: ^9.1.2
+  in_app_review: ^2.0.6
+  introduction_screen: ^3.1.11
+  json_annotation: ^4.8.1
+  json_serializable: ^6.7.1
   # lan_scanner: ^3.5.0
   meta: ^1.8.0
   mocktail: ^0.3.0
   modal_bottom_sheet: ^3.0.0-pre
-  network_info_plus: ^3.0.1
-  network_tools: ^3.0.0+2
-  package_info_plus: ^3.0.2
-  path: ^1.8.2
-  path_provider: ^2.0.12
-  permission_handler: ^10.2.0
-  sentry_flutter: ^6.20.1
-  url_launcher: ^6.1.9
-  wake_on_lan: ^3.1.0+6
-  wiredash: ^1.5.0
+  network_info_plus: ^4.0.1
+  network_tools: ^3.2.1
+  package_info_plus: ^4.0.2
+  path_provider: ^2.0.15
+  permission_handler: ^10.4.3
+  sentry_flutter: ^7.9.0
+  url_launcher: ^6.1.12
+  wake_on_lan: ^3.2.0+1
 
 dev_dependencies:
-  bloc_test: ^9.1.0
-  build_runner: ^2.2.1
-  dart_code_metrics: ^4.19.1
+  bloc_test: ^9.1.3
+  build_runner: ^2.4.6
   flutter_test:
     sdk: flutter
-  icons_launcher: ^2.0.5
   import_sorter: ^4.6.0
-  mockingjay: ^0.3.0
-  very_good_analysis: ^3.1.0
+  very_good_analysis: ^5.0.0+1
 
 flutter:
   uses-material-design: true
