@@ -1,35 +1,25 @@
 // Package imports:
+import 'package:carrier_info/carrier_info.dart';
 import 'package:equatable/equatable.dart';
 
 class CarrierInfoModel extends Equatable {
   const CarrierInfoModel({
-    required this.allowsVOIP,
-    this.carrierName,
-    this.isoCountryCode,
-    this.mobileCountryCode,
-    this.mobileNetworkCode,
-    this.networkGeneration,
-    this.radioType,
+    required this.androidCarrierData,
+    required this.iosCarrierData,
   });
 
-  final bool allowsVOIP;
-  final String? carrierName;
-  final String? isoCountryCode;
-  final String? mobileCountryCode;
-  final String? mobileNetworkCode;
-  final String? networkGeneration;
-  final String? radioType;
+  final AndroidCarrierData? androidCarrierData;
+  final IosCarrierData? iosCarrierData;
 
-  bool get isCarrierConnected => carrierName != null;
+  bool get isCarrierConnected =>
+      androidCarrierData?.isSmsCapable ??
+      false ||
+          iosCarrierData!.carrierData
+              .any((element) => element.mobileNetworkCode != '');
 
   @override
   List<Object?> get props => [
-        allowsVOIP,
-        carrierName,
-        isoCountryCode,
-        mobileCountryCode,
-        mobileNetworkCode,
-        networkGeneration,
-        radioType,
+        androidCarrierData,
+        iosCarrierData,
       ];
 }
