@@ -3,18 +3,16 @@ import 'package:carrier_info/carrier_info.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 // Project imports:
-import 'package:network_arch/network_status/models/models.dart';
+import 'package:network_arch/network_status/network_status.dart';
 
 class CarrierDataProvider {
-  Future<CarrierInfoModel> getCellularData() async {
+  Future<CarrierInfoModel?> getCellularData() async {
+    final android = await CarrierInfo.getAndroidInfo();
+    final ios = await CarrierInfo.getIosInfo();
+
     return CarrierInfoModel(
-      allowsVOIP: await CarrierInfo.allowsVOIP,
-      carrierName: await CarrierInfo.carrierName,
-      isoCountryCode: await CarrierInfo.isoCountryCode,
-      mobileCountryCode: await CarrierInfo.mobileCountryCode,
-      mobileNetworkCode: await CarrierInfo.mobileNetworkCode,
-      networkGeneration: await CarrierInfo.networkGeneration,
-      radioType: await CarrierInfo.radioType,
+      androidCarrierData: android,
+      iosCarrierData: ios,
     );
   }
 
