@@ -56,7 +56,11 @@ class CarrierStatusCard extends StatelessWidget {
                           const Text('Network Generation'),
                           const Spacer(),
                           if (state.carrierStatus == NetworkStatus.success)
-                            Text(state.carrierInfo!.networkGeneration ?? 'N/A')
+                            Text(
+                              state.carrierInfo?.androidCarrierData
+                                      ?.telephonyInfo.first.networkGeneration ??
+                                  'N/A',
+                            )
                           else if (state.carrierStatus ==
                               NetworkStatus.permissionIssue)
                             const Text('N/A')
@@ -107,7 +111,9 @@ class CarrierStatusCard extends StatelessWidget {
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
                           Text(
-                            state.carrierInfo?.carrierName ?? 'N/A',
+                            state.carrierInfo?.iosCarrierData?.carrierData.first
+                                    .carrierName ??
+                                'N/A',
                             style: const TextStyle(
                               fontWeight: FontWeight.bold,
                               fontSize: 20,
@@ -127,7 +133,14 @@ class CarrierStatusCard extends StatelessWidget {
                           const Text('Network Generation'),
                           const Spacer(),
                           if (state.carrierStatus == NetworkStatus.success)
-                            Text(state.carrierInfo!.networkGeneration ?? 'N/A')
+                            Text(
+                              state
+                                      .carrierInfo
+                                      ?.iosCarrierData
+                                      ?.carrierRadioAccessTechnologyTypeList
+                                      .first ??
+                                  'N/A',
+                            )
                           else if (state.carrierStatus ==
                               NetworkStatus.permissionIssue)
                             const Text('N/A')
