@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 // Project imports:
-import 'package:network_arch/theme/themes.dart';
+import 'package:network_arch/shared/shared.dart';
 
 class HexBytesViewer extends StatelessWidget {
   const HexBytesViewer({
@@ -28,63 +28,43 @@ class HexBytesViewer extends StatelessWidget {
           ),
           const SizedBox(height: 10),
         ],
-        Card(
+        DataCard(
           margin: EdgeInsets.zero,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(10),
-          ),
-          color: Theme.of(context).platform != TargetPlatform.iOS
-              ? Color.alphaBlend(
-                  Theme.of(context).colorScheme.primary.withOpacity(0.1),
-                  Theme.of(context).colorScheme.surfaceVariant,
-                )
-              : Themes.iOSOnboardingBgColor.resolveFrom(context),
-          elevation: 0,
-          child: Padding(
-            padding: const EdgeInsets.all(8),
-            child: Column(
-              children: [
-                GridView.count(
-                  physics: const NeverScrollableScrollPhysics(),
-                  crossAxisCount: 16,
-                  shrinkWrap: true,
-                  padding: EdgeInsets.zero,
-                  children: [
-                    for (final byte in bytes)
+          child: Column(
+            children: [
+              GridView.count(
+                physics: const NeverScrollableScrollPhysics(),
+                crossAxisCount: 16,
+                shrinkWrap: true,
+                padding: EdgeInsets.zero,
+                children: [
+                  for (final byte in bytes)
+                    Text(
+                      byte.toRadixString(16).padLeft(2, '0'),
+                      style: GoogleFonts.sourceCodePro(),
+                    ),
+                ],
+              ),
+              GridView.count(
+                physics: const NeverScrollableScrollPhysics(),
+                crossAxisCount: 16,
+                shrinkWrap: true,
+                padding: EdgeInsets.zero,
+                children: [
+                  for (final byte in bytes)
+                    if (byte < 128)
                       Text(
-                        byte.toRadixString(16).padLeft(2, '0'),
-                        style: GoogleFonts.sourceCodePro(
-                          // On android, this makes no difference
-                          color: Themes.iOStextColor.resolveFrom(context),
-                        ),
+                        String.fromCharCode(byte),
+                        style: GoogleFonts.sourceCodePro(),
+                      )
+                    else
+                      Text(
+                        '.',
+                        style: GoogleFonts.sourceCodePro(),
                       ),
-                  ],
-                ),
-                GridView.count(
-                  physics: const NeverScrollableScrollPhysics(),
-                  crossAxisCount: 16,
-                  shrinkWrap: true,
-                  padding: EdgeInsets.zero,
-                  children: [
-                    for (final byte in bytes)
-                      if (byte < 128)
-                        Text(
-                          String.fromCharCode(byte),
-                          style: GoogleFonts.sourceCodePro(
-                            color: Themes.iOStextColor.resolveFrom(context),
-                          ),
-                        )
-                      else
-                        Text(
-                          '.',
-                          style: GoogleFonts.sourceCodePro(
-                            color: Themes.iOStextColor.resolveFrom(context),
-                          ),
-                        ),
-                  ],
-                ),
-              ],
-            ),
+                ],
+              ),
+            ],
           ),
         ),
       ],
