@@ -10,7 +10,6 @@ import 'package:hive/hive.dart';
 // Project imports:
 import 'package:network_arch/constants.dart';
 import 'package:network_arch/shared/shared.dart';
-import 'package:network_arch/theme/theme.dart';
 import 'package:network_arch/utils/utils.dart';
 import 'package:network_arch/whois/whois.dart';
 
@@ -96,8 +95,6 @@ class _WhoisViewState extends State<WhoisView> {
   }
 
   Widget _buildBody(BuildContext context) {
-    final isIOS = Theme.of(context).platform == TargetPlatform.iOS;
-
     return ContentListView(
       usePaddingOniOS: true,
       children: [
@@ -114,17 +111,10 @@ class _WhoisViewState extends State<WhoisView> {
         BlocBuilder<WhoisBloc, WhoisState>(
           builder: (context, state) {
             if (state is WhoisLoadSuccess) {
-              return SlideInUp(
-                child: FadeIn(
-                  child: DataCard(
-                    child: Text(
-                      state.response,
-                      style: TextStyle(
-                        color: isIOS
-                            ? Themes.iOStextColor.resolveFrom(context)
-                            : null,
-                      ),
-                    ),
+              return FadeIn(
+                child: DataCard(
+                  child: Text(
+                    state.response,
                   ),
                 ),
               );
